Refactor BookShopsApiController.cs
